tif_to_zarr/requirements.txt
xarray
zarr
geopandas
numpy
netcdf4
rasterio
requests
dask
tqdm
rioxarray
